operator/requirements.txt
kopf==1.36.2
PyYAML
kubernetes